Remove unnecessary __tests__ folders
