+++ packages/ketcher-polymer-editor-react/src/components/shared/actionButton/ActionButton.test.tsx
@@ -17,7 +17,7 @@
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 
-import { ActionButton } from '..'
+import { ActionButton } from '.'
 
 const mockClickHandler = jest.fn()
 const MOCK_LABEL = 'Click Me!'
